+++ lib/view/pages/user/login/login_page.dart
@@ -25,7 +25,6 @@ class LoginPage extends StatelessWidget {
                 Text("로그인", style: textTheme().headline1),
                 SizedBox(height: gap_xl),
                 CustomFormLogin(),
-                _buildLoginButton(context),
                 SizedBox(height: gap_xl),
                 TextButton(
                   onPressed: () {
@@ -36,30 +35,6 @@ class LoginPage extends StatelessWidget {
               ],
             ),
           ],
-        ),
-      ),
-    );
-  }
-
-  Padding _buildLoginButton(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: gap_m),
-      child: Container(
-        width: double.infinity,
-        height: 40,
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(10),
-          border: Border.all(color: kMainColor),
-          color: kMainColor,
-        ),
-        child: TextButton(
-          onPressed: () {
-            Navigator.push(context, MaterialPageRoute(builder: (context) => MainPage()));
-          },
-          child: Text(
-            '로그인',
-            style: TextStyle(color: Colors.white, fontSize: 14, height: 1.0),
-          ),
         ),
       ),
     );
